feat(mantine): enable table to have multiple layouts
