+++ packages/mantine/src/components/table/TableLoading.tsx
@@ -1,0 +1,8 @@
+import {Skeleton, SkeletonProps} from '@mantine/core';
+import {FunctionComponent} from 'react';
+
+export const TableLoading: FunctionComponent<SkeletonProps> = (props) => (
+    <Skeleton style={{display: 'inline-block'}} {...props} sx={!props.visible ? {borderRadius: 0} : undefined}>
+        {props.children}
+    </Skeleton>
+);
