+++ packages/mantine/src/components/table/index.ts
@@ -1,3 +1,4 @@
 export * from './Table';
 export {useTable} from './TableContext';
 export {type onTableChangeEvent, type InitialTableState, type TableState, type TableProps} from './Table.types';
+export {TableLayouts} from './layouts/TableLayouts';
