+++ packages/mantine/src/components/table/layouts/TableLayouts.tsx
@@ -1,0 +1,5 @@
+import {RowLayout} from './RowLayout';
+
+export const TableLayouts = {
+    Rows: RowLayout,
+};
