+++ packages/website/src/pages/layout/Table.tsx
@@ -1,10 +1,10 @@
 import {TableMetadata} from '@coveord/plasma-components-props-analyzer';
 import TableDemo from '@examples/layout/Table/Table.demo?demo';
-import TableClientSideDemo from '@examples/layout/Table/TableClientSide.demo?demo';
-import TableConsumerDemo from '@examples/layout/Table/TableConsumer.demo?demo';
-import TableDisableRowSelection from '@examples/layout/Table/TableDisabledRowSelection.demo?demo';
-import TableEmptyStateDemo from '@examples/layout/Table/TableEmptyState.demo?demo';
-import TableMultiSelectionDemo from '@examples/layout/Table/TableMultiSelection.demo?demo';
+// import TableClientSideDemo from '@examples/layout/Table/TableClientSide.demo?demo';
+// import TableConsumerDemo from '@examples/layout/Table/TableConsumer.demo?demo';
+// import TableDisableRowSelection from '@examples/layout/Table/TableDisabledRowSelection.demo?demo';
+// import TableEmptyStateDemo from '@examples/layout/Table/TableEmptyState.demo?demo';
+// import TableMultiSelectionDemo from '@examples/layout/Table/TableMultiSelection.demo?demo';
 
 import {PageLayout} from '../../building-blocs/PageLayout';
 
@@ -17,15 +17,15 @@ const DemoPage = () => (
         id="Table"
         propsMetadata={TableMetadata}
         demo={<TableDemo noPadding layout="vertical" />}
-        examples={{
-            multiSelect: <TableMultiSelectionDemo noPadding title="Table with bulk selection of rows" />,
-            disableRowSelection: <TableDisableRowSelection noPadding title="Table with disabled row selection" />,
-            clientSide: (
-                <TableClientSideDemo noPadding title="Table with client side pagination, sorting, and filtering" />
-            ),
-            emptyState: <TableEmptyStateDemo noPadding title="Table with empty states" />,
-            consumer: <TableConsumerDemo noPadding title="Table with a child component using the hook to re-fetch" />,
-        }}
+        // examples={{
+        //     multiSelect: <TableMultiSelectionDemo noPadding title="Table with bulk selection of rows" />,
+        //     disableRowSelection: <TableDisableRowSelection noPadding title="Table with disabled row selection" />,
+        //     clientSide: (
+        //         <TableClientSideDemo noPadding title="Table with client side pagination, sorting, and filtering" />
+        //     ),
+        //     emptyState: <TableEmptyStateDemo noPadding title="Table with empty states" />,
+        //     consumer: <TableConsumerDemo noPadding title="Table with a child component using the hook to re-fetch" />,
+        // }}
     />
 );
